chore: Change Object count type from i32 to i64
This ensures a more consistent conversion between the individual
representations of an object. In case of doubt, the maximum number of
sub-resources could be limited with a constraint on the column in the
database.

+++ src/database/schema.sql
@@ -148,7 +148,7 @@ CREATE TABLE IF NOT EXISTS objects (
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     created_by UUID NOT NULL REFERENCES users(id),
     content_len BIGINT NOT NULL DEFAULT 0,
-    count INT NOT NULL DEFAULT 0,
+    count BIGINT NOT NULL DEFAULT 0,
     key_values JSONB NOT NULL,
     object_status "ObjectStatus" NOT NULL DEFAULT 'INITIALIZING',
     data_class "DataClass" NOT NULL DEFAULT 'PRIVATE',
